requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-filer==1.4.4
django-js-asset==1.2.2
django-mptt==0.9.1
django-oauth-toolkit==1.2.0
django-polymorphic==2.0.3
djangorestframework==3.9.2
easy-thumbnails==2.6
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
requests==2.21.0
shortuuid==0.5.0
Unidecode==1.0.23
urllib3==1.24.1
